E2E credentials (#2916)
* improve credentials e2e tests

* add tests

* fix e2e

* fix: front adaptation to operatorMiddleware

---------

+++ api/src/pdc/services/auth/actions/DeleteCredentialsAction.ts
@@ -1,14 +1,21 @@
 import { handler } from "@/ilos/common/Decorators.ts";
+import { NotFoundException } from "@/ilos/common/index.ts";
 import { Action as AbstractAction } from "@/ilos/core/index.ts";
-import { copyFromContextMiddleware } from "@/pdc/providers/middleware/middlewares.ts";
+import { enforceOperatorMiddleware, hasPermissionByScopeMiddleware } from "@/pdc/providers/middleware/middlewares.ts";
 import { DexClient } from "@/pdc/services/auth/providers/DexClient.ts";
 import { DeleteCredentialsParams, DeleteCredentialsResult } from "../dto/Credentials.ts";
+import { OperatorRepository } from "../providers/OperatorRepository.ts";
 
 @handler({
   service: "auth",
   method: "deleteCredentials",
   middlewares: [
-    copyFromContextMiddleware(`call.user.operator_id`, "operator_id", false),
+    enforceOperatorMiddleware(),
+    hasPermissionByScopeMiddleware("registry.credentials.delete", [
+      "operator.credentials.delete",
+      "call.user.operator_id",
+      "operator_id",
+    ]),
     ["validate", DeleteCredentialsParams],
   ],
   apiRoute: {
@@ -18,11 +25,15 @@ import { DeleteCredentialsParams, DeleteCredentialsResult } from "../dto/Credent
   },
 })
 export class DeleteCredentialsAction extends AbstractAction {
-  constructor(private dexClient: DexClient) {
+  constructor(private dexClient: DexClient, private operatorRepository: OperatorRepository) {
     super();
   }
 
   protected override async handle(params: DeleteCredentialsParams): Promise<DeleteCredentialsResult> {
+    if (!params.operator_id || !(await this.operatorRepository.exists(params.operator_id))) {
+      throw new NotFoundException(`Operator with ID ${params.operator_id} does not exist`);
+    }
+
     return this.dexClient.deleteByOperator(params.operator_id, params.token_id);
   }
 }
